Began resolving compile errors: 526 remain.

+++ Game/Misc/Wires_Vending.cs
@@ -23,7 +23,7 @@ namespace Somnium.Game {
 
 			V = this.holder;
 
-			switch ((double?)( index )) {
+			switch ((int?)( index )) {
 				case 1:
 					((dynamic)V).shoot_inventory = !mended;
 					break;
